Adicionado Flag Debug , ajustado tipos de variáveis das funções de cálculo de erro relativo e absoluto. Complementado alguns comentários

+++ Exercicios/Ep01/floatType.c
@@ -79,13 +79,13 @@ int AlmostEqualUlps(Float_t A, Float_t B, int maxULPs)
 }
 
 //Retorna o Erro Absoluto Real
-double erro_abs(double a, double b)
+float erro_abs(float a, float b)
 {
   return b - a;
 }
 
 //Retorna o Erro Relativo Real
-double erro_rel(double a, double b)
+float erro_rel(float a, float b)
 {
    return (b -a) / a;
 }
